+++ Source/Krypton Components/Krypton.Toolkit/Controls Toolkit/KryptonManager.cs
@@ -25,7 +25,7 @@ namespace Krypton.Toolkit
         #region Static Fields
         // Initialize the global state
         private static bool _globalApplyToolstrips = true;
-        private static bool _globalAllowFormChrome = true;
+        private static bool _globalUseThemeFormChromeBorderWidth = true;
         internal static bool _globalUseKryptonFileDialogs = true;
         private static Font? _baseFont;
 
@@ -141,11 +141,11 @@ namespace Krypton.Toolkit
         public static event EventHandler? GlobalPaletteChanged;
 
         /// <summary>
-        /// Occurs when the AllowFormChrome property changes.
+        /// Occurs when the UseThemeFormChromeBorderWidth property changes.
         /// </summary>
         [Category(@"Property Changed")]
-        [Description(@"Occurs when the value of the GlobalAllowFormChrome property is changed.")]
-        public static event EventHandler? GlobalAllowFormChromeChanged;
+        [Description(@"Occurs when the value of the GlobalUseThemeFormChromeBorderWidth property is changed.")]
+        public static event EventHandler? GlobalUseThemeFormChromeBorderWidthChanged;
         #endregion
 
         #region Identity
@@ -203,14 +203,16 @@ namespace Krypton.Toolkit
         /// <summary>
         /// Have any of the global values been modified
         /// </summary>
+        [Browsable(false)]
         [EditorBrowsable(EditorBrowsableState.Never)]
         [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
         public bool IsDefault => !(ShouldSerializeGlobalCustomPalette() ||
                                    ShouldSerializeGlobalApplyToolstrips() ||
-                                   ShouldSerializeGlobalAllowFormChrome() ||
+                                   ShouldSerializeGlobalUseThemeFormChromeBorderWidth() ||
                                    ShouldSerializeToolkitStrings() ||
                                    ShouldSerializeUseKryptonFileDialogs() ||
-                                   ShouldSerializeBaseFont()
+                                   ShouldSerializeBaseFont() ||
+                                   ShouldSerializeGlobalPaletteMode()
                                    );
 
         /// <summary>
@@ -220,10 +222,11 @@ namespace Krypton.Toolkit
         {
             ResetGlobalCustomPalette();
             ResetGlobalApplyToolstrips();
-            ResetGlobalAllowFormChrome();
+            ResetGlobalUseThemeFormChromeBorderWidth();
             ResetToolkitStrings();
             ResetUseKryptonFileDialogs();
             ResetBaseFont();
+            ResetGlobalPaletteMode();
         }
 
         /// <summary>
@@ -325,13 +328,12 @@ namespace Krypton.Toolkit
             }
         }
 
-        internal void ResetBaseFont()
+        private void ResetBaseFont()
         {
             _baseFont = null;
             CurrentGlobalPalette.ResetBaseFont();
         }
-
-        internal bool ShouldSerializeBaseFont() => _baseFont != null;
+        private bool ShouldSerializeBaseFont() => _baseFont != null;
 
         /// <summary>
         /// Gets or sets a value indicating if the palette colors are applied to the tool-strips.
@@ -362,18 +364,18 @@ namespace Krypton.Toolkit
 
 
         /// <summary>
-        /// Gets or sets a value indicating if KryptonForm instances are allowed to show custom chrome.
+        /// Gets or sets a value indicating if KryptonForm instances are allowed to UseThemeFormChromeBorderWidth.
         /// </summary>
         [Category(@"Visuals")]
-        [Description(@"Should KryptonForm instances be allowed to show custom chrome.")]
+        [Description(@"Should KryptonForm instances be allowed to UseThemeFormChromeBorderWidth.")]
         [DefaultValue(true)]
-        public bool GlobalAllowFormChrome
+        public bool GlobalUseThemeFormChromeBorderWidth
         {
-            get => AllowFormChrome;
-            set => AllowFormChrome = value;
+            get => UseThemeFormChromeBorderWidth;
+            set => UseThemeFormChromeBorderWidth = value;
         }
-        private bool ShouldSerializeGlobalAllowFormChrome() => !GlobalAllowFormChrome;
-        private void ResetGlobalAllowFormChrome() => GlobalAllowFormChrome = true;
+        private bool ShouldSerializeGlobalUseThemeFormChromeBorderWidth() => !GlobalUseThemeFormChromeBorderWidth;
+        private void ResetGlobalUseThemeFormChromeBorderWidth() => GlobalUseThemeFormChromeBorderWidth = true;
 
         /// <summary>Gets the toolkit strings that can be localised.</summary>
         [Category(@"Data")]
@@ -425,24 +427,24 @@ namespace Krypton.Toolkit
         }
         #endregion
 
-        #region Static AllowFormChrome
+        #region Static UseThemeFormChromeBorderWidth
         /// <summary>
         /// Gets and sets the global flag that decides if form chrome should be customized.
         /// </summary>
-        public static bool AllowFormChrome
+        public static bool UseThemeFormChromeBorderWidth
         {
-            get => _globalAllowFormChrome;
+            get => _globalUseThemeFormChromeBorderWidth;
 
             set
             {
                 // Only interested if the value changes
-                if (_globalAllowFormChrome != value)
+                if (_globalUseThemeFormChromeBorderWidth != value)
                 {
                     // Use new value
-                    _globalAllowFormChrome = value;
+                    _globalUseThemeFormChromeBorderWidth = value;
 
                     // Fire change event
-                    OnGlobalAllowFormChromeChanged(EventArgs.Empty);
+                    OnGlobalUseThemeFormChromeBorderWidthChanged(EventArgs.Empty);
                 }
             }
         }
@@ -1003,7 +1005,7 @@ namespace Krypton.Toolkit
             }
         }
 
-        private static void OnGlobalAllowFormChromeChanged(EventArgs e) => GlobalAllowFormChromeChanged?.Invoke(null, e);
+        private static void OnGlobalUseThemeFormChromeBorderWidthChanged(EventArgs e) => GlobalUseThemeFormChromeBorderWidthChanged?.Invoke(null, e);
 
         private static void OnGlobalPaletteChanged(EventArgs e)
         {